Crouch  bug fix
players now get their own mesh and the crouch bug does not happen visually. The colision shape bug probably still happnes.. ok fixed that too! In the future, instead of changing the scale of the mesh, I should just have a crouching and standing mesh so instead of each player getting a resource, all players have two. This is assuming I don't do dynamic states or hights for the meshes.