Launcher code change & PID tuner removal
Removed the PID tuner. If we need it back, get it from git where I deleted it right here.

trying to make the launcher only shoot when it should, also made code more readable

Removed a whole bunch of other testing junk that I used to learn stuff that is either commented out or doesn't do anything.

This is the first step towards me seperating the functionalities from the gamepad inputs so I can use my teleop code in auto and vice versa.