Pedro & Limlight work together pls
Reversed Pedro, working on auto

Pedro forward must = robot forward to work with the limelight. This will break my old auto but that's ok.

Ok so here's the state of the drivetrain: the leftFront wheel in the config and in the code is acctually the right back wheel becuase I spun it around. fixing this now before I go..