tweaks
not sure if this is going to work or not. if you average out the undershoot that it was and the overshoot that it is, you get average on target even when you're not on target. this might be a poor idea, I should probably just tune the pid. mabye check if the past 5 are all nearby individually instead of the average, that could work!