Panels Visualisation
This should allow me to see the motor's speed in panels which will make it very easy to tune the launcher PID. I should experiment with a power supply for the robot instead of just battery that changes voltage so we can test individual voltages. Seeing the speed over time and when a ball enters will make this way better!