Work needed
working on it. In the future I need to make it so I can still drive during this testing teleop. Probably need to rely more on the teleop class instead of having duplicate features!

Also I added lines to show where the thing is targetting.

I should have PID variables for the left motor, right motor, and configurable for the launch angle.